requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
altair>=5.0.0
requests>=2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
